Remove "can_handle_data" from the CAN module turning it finally over into a low-level hardware driver and concentrating all business logic in the module "can_handler".
